LayoutTests/svg/animations/animations-lowPowerMode-throttling-expected.txt

Tests that SVG animations are properly throttled in low power mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.svgAnimationsInterval(svgElement) is 1/60.
internals.setLowPowerModeEnabled(true)
PASS internals.svgAnimationsInterval(svgElement) is 1/30.
internals.setLowPowerModeEnabled(false)
PASS internals.svgAnimationsInterval(svgElement) is 1/60.
PASS successfullyParsed is true

TEST COMPLETE
